🖋 rename : ViewConst 패키지와 이름 변경
